Clean /auth URLs for timing measurements
`/auth` URLs sometimes include the app's ID. In once case, it's at the end and
we can just drop it. In another, we have to pull it from the middle of the
string. I replace it with `APP_ID` and make sure the path is always lower case
so we can tell that's a special segment.