fix a bug in profileManager.authGet
Because of #704, we had lots of profiles with missing auth.apps.<app-id>
objects. Previously, if an app-specfic auth was missing, authGet would just
return the whole auth object, but this cause a problems for auth=true endpoints
and posting because it could potentially use the wrong tokens.

This fixes that, and also adds more verbose and throrough error checking +
logging to places that this will cause problems.